Moving DetailsNameField and associated components to product-editor package
