+++ packages/js/product-editor/src/components/edit-product-link-modal/index.ts
