+++ packages/js/product-editor/src/components/edit-product-link-modal/test/edit-product-link-modal.test.tsx
@@ -4,6 +4,7 @@
 import { render } from '@testing-library/react';
 import { Product } from '@woocommerce/data';
 import userEvent from '@testing-library/user-event';
+import { createElement } from '@wordpress/element';
 
 /**
  * Internal dependencies
