+++ packages/js/product-editor/src/components/field-details-name/index.ts
@@ -1,0 +1,1 @@
+export * from './field-details-name';
