+++ packages/js/product-editor/src/constants.ts
@@ -1,3 +1,13 @@
 export const NUMBERS_AND_ALLOWED_CHARS = '[^-0-9%s1%s2]';
 export const NUMBERS_AND_DECIMAL_SEPARATOR = '[^-\\d\\%s]+';
 export const ONLY_ONE_DECIMAL_SEPARATOR = '[%s](?=%s*[%s])';
+
+// This should never be a real slug value of any existing shipping class
+export const ADD_NEW_SHIPPING_CLASS_OPTION_VALUE =
+	'__ADD_NEW_SHIPPING_CLASS_OPTION__';
+export const UNCATEGORIZED_CATEGORY_SLUG = 'uncategorized';
+export const PRODUCT_VARIATION_TITLE_LIMIT = 32;
+export const STANDARD_RATE_TAX_CLASS_SLUG = 'standard';
+
+export const PLUGIN_ID = 'woocommerce';
+export const PRODUCT_DETAILS_SLUG = 'product-details';
