+++ packages/js/product-editor/src/style.scss
@@ -1,1 +1,2 @@
 @import 'components/product-section-layout/style.scss';
+@import 'components/edit-product-link-modal/style.scss';
