+++ plugins/woocommerce-admin/client/products/fills/details-section/index.ts
@@ -1,4 +1,3 @@
-export * from './details-field-name';
 export * from './details-field-categories';
 export * from './details-field-feature';
 export * from './details-field-summary';
